# backend/src/Api/Controllers/FamiliesController.cs
using FamilyVeda.Application.Common;
using FamilyVeda.Application.Families;
using FluentValidation;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace FamilyVeda.Api.Controllers;

[Route("api/v1/families")]
[Authorize(Policy = "FamilyUser")]
public sealed class FamiliesController(IFamilyService familyService) : ApiControllerBase
{
    [HttpGet("me")]
    public async Task<ActionResult<FamilyDto>> GetMine(CancellationToken cancellationToken)
    {
        var family = await familyService.GetMineAsync(cancellationToken);
        return family is null ? NotFound() : Ok(family);
    }

    [HttpPost]
    public async Task<ActionResult<FamilyDto>> Create(CreateFamilyRequest request, IValidator<CreateFamilyRequest> validator, CancellationToken cancellationToken)
    {
        await ValidateAsync(validator, request, cancellationToken);
        var family = await familyService.CreateAsync(request, cancellationToken);
        return CreatedAtAction(nameof(GetMine), family);
    }

    [HttpPut("{familyId:guid}")]
    public async Task<ActionResult<FamilyDto>> Update(Guid familyId, UpdateFamilyRequest request, CancellationToken cancellationToken) =>
        Ok(await familyService.UpdateAsync(familyId, request, cancellationToken));

    [HttpGet("{familyId:guid}/members")]
    public async Task<ActionResult<PagedResult<MemberDto>>> GetMembers(Guid familyId, int page = 1, int pageSize = 20, string? search = null, CancellationToken cancellationToken = default) =>
        Ok(await familyService.GetMembersAsync(familyId, page, pageSize, search, cancellationToken));

    [HttpPost("{familyId:guid}/members")]
    public async Task<ActionResult<MemberDto>> AddMember(Guid familyId, CreateMemberRequest request, IValidator<CreateMemberRequest> validator, CancellationToken cancellationToken)
    {
        await ValidateAsync(validator, request, cancellationToken);
        var member = await familyService.AddMemberAsync(familyId, request, cancellationToken);
        return Created($"/api/v1/members/{member.Id}", member);
    }

    [HttpPost("{familyId:guid}/invitations")]
    public async Task<ActionResult<FamilyInvitationDto>> CreateInvitation(Guid familyId, CreateFamilyInvitationRequest request, CancellationToken cancellationToken)
    {
        var invitation = await familyService.CreateInvitationAsync(familyId, request, cancellationToken);
        return Created($"/api/v1/families/{familyId}/invitations/{invitation.Id}", invitation);
    }

    [HttpPost("invitations/accept")]
    public async Task<ActionResult<MemberDto>> AcceptInvitation(AcceptFamilyInvitationRequest request, CancellationToken cancellationToken) =>
        Ok(await familyService.AcceptInvitationAsync(request, cancellationToken));
}