# backend/src/Api/Controllers/MembersController.cs
using FamilyVeda.Application.Families;
using FamilyVeda.Domain.Common;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace FamilyVeda.Api.Controllers;

[Route("api/v1/members")]
[Authorize(Policy = "FamilyUser")]
public sealed class MembersController(IFamilyService familyService) : ApiControllerBase
{
    [HttpGet("me")]
    public async Task<ActionResult<MemberDto>> GetMine(CancellationToken cancellationToken) =>
        Ok(await familyService.GetMyMemberAsync(cancellationToken));

    [HttpGet("{memberId:guid}")]
    public async Task<ActionResult<MemberDto>> Get(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await familyService.GetMemberAsync(memberId, cancellationToken));

    [HttpPut("{memberId:guid}")]
    public async Task<ActionResult<MemberDto>> Update(Guid memberId, UpdateMemberRequest request, CancellationToken cancellationToken) =>
        Ok(await familyService.UpdateMemberAsync(memberId, request, cancellationToken));

    [HttpDelete("{memberId:guid}")]
    public async Task<ActionResult<object>> Delete(Guid memberId, CancellationToken cancellationToken)
    {
        await familyService.DeleteMemberAsync(memberId, cancellationToken);
        return NoContent();
    }

    [HttpGet("{memberId:guid}/relationships")]
    public async Task<ActionResult<IReadOnlyList<RelationshipDto>>> GetRelationships(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await familyService.GetRelationshipsAsync(memberId, cancellationToken));

    [HttpPost("{memberId:guid}/relationships")]
    public async Task<ActionResult<RelationshipDto>> AddRelationship(Guid memberId, CreateRelationshipRequest request, CancellationToken cancellationToken)
    {
        var relationship = await familyService.AddRelationshipAsync(memberId, request, cancellationToken);
        return Created($"/api/v1/members/{memberId}/relationships/{relationship.Id}", relationship);
    }

    [HttpGet("{memberId:guid}/consents")]
    public async Task<ActionResult<IReadOnlyList<ConsentDto>>> GetConsents(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await familyService.GetConsentsAsync(memberId, cancellationToken));

    [HttpPut("{memberId:guid}/consents/{category}")]
    public async Task<ActionResult<ConsentDto>> UpdateConsent(Guid memberId, ConsentCategory category, UpdateConsentRequest request, CancellationToken cancellationToken) =>
        Ok(await familyService.UpdateConsentAsync(memberId, category, request, cancellationToken));

    [HttpPost("{memberId:guid}/consents/reaffirm")]
    public async Task<ActionResult<IReadOnlyList<ConsentDto>>> Reaffirm(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await familyService.ReaffirmConsentsAsync(memberId, cancellationToken));
}