# docs/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: Family Veda API
  version: v1
paths:
...
  "/api/v1/families":
    post:
      tags:
      - Families
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FamilyDto"
  "/api/v1/families/{familyId}":
    put:
      tags:
      - Families
      parameters:
      - name: familyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateFamilyRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/UpdateFamilyRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UpdateFamilyRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FamilyDto"
  "/api/v1/families/{familyId}/members":
    get:
      tags:
      - Families
      parameters:
      - name: familyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: search
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemberDtoPagedResult"
    post:
      tags:
      - Families
      parameters:
      - name: familyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateMemberRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateMemberRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateMemberRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemberDto"
  "/api/v1/families/{familyId}/invitations":
    post:
      tags:
      - Families
      parameters:
      - name: familyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyInvitationRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyInvitationRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateFamilyInvitationRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FamilyInvitationDto"
  "/api/v1/families/invitations/accept":
    post:
      tags:
      - Families
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AcceptFamilyInvitationRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/AcceptFamilyInvitationRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/AcceptFamilyInvitationRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemberDto"
...
  "/api/v1/members/{memberId}":
    get:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemberDto"
    put:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateMemberRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/UpdateMemberRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UpdateMemberRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemberDto"
    delete:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  "/api/v1/members/{memberId}/relationships":
    get:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/RelationshipDto"
    post:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateRelationshipRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateRelationshipRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateRelationshipRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RelationshipDto"
  "/api/v1/members/{memberId}/consents":
    get:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ConsentDto"
  "/api/v1/members/{memberId}/consents/{category}":
    put:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: category
        in: path
        required: true
        schema:
          "$ref": "#/components/schemas/ConsentCategory"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateConsentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/UpdateConsentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UpdateConsentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ConsentDto"
  "/api/v1/members/{memberId}/consents/reaffirm":
    post:
      tags:
      - Members
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ConsentDto"
...
components:
  schemas:
    AcceptFamilyInvitationRequest:
      type: object
      properties:
        token:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
      additionalProperties: false
...
    ConsentCategory:
      enum:
      - HereditaryFlags
      - VitalsSummary
      - Conditions
      type: string
    ConsentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        category:
          "$ref": "#/components/schemas/ConsentCategory"
        status:
          "$ref": "#/components/schemas/ConsentStatus"
        grantedByGuardian:
          type: boolean
      additionalProperties: false
    ConsentStatus:
      enum:
      - NotSet
      - Granted
      - Revoked
      - PendingReaffirmation
      type: string
...
    CreateFamilyInvitationRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
      additionalProperties: false
    CreateFamilyRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
      additionalProperties: false
    CreateMemberRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
        role:
          "$ref": "#/components/schemas/FamilyRole"
        userId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
    CreateRelationshipRequest:
      type: object
      properties:
        relatedMemberId:
          type: string
          format: uuid
        relationshipType:
          type: string
          nullable: true
        isBiological:
          type: boolean
      additionalProperties: false
...
    FamilyDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        members:
          type: array
          items:
            "$ref": "#/components/schemas/MemberDto"
          nullable: true
      additionalProperties: false
    FamilyInvitationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
      additionalProperties: false
    FamilyRole:
      enum:
      - Head
      - AdultMember
      - MinorMember
      type: string
...
    MemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
        role:
          "$ref": "#/components/schemas/FamilyRole"
      additionalProperties: false
    MemberDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/MemberDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
    RelationshipDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        relatedMemberId:
          type: string
          format: uuid
        relationshipType:
          type: string
          nullable: true
        isBiological:
          type: boolean
      additionalProperties: false
...
    UpdateConsentRequest:
      type: object
      properties:
        status:
          "$ref": "#/components/schemas/ConsentStatus"
      additionalProperties: false
    UpdateFamilyRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
      additionalProperties: false
    UpdateMemberRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
        role:
          "$ref": "#/components/schemas/FamilyRole"
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- Bearer: []